requirements.txt
torch
torchaudio
numpy
scipy
komm
scikit-commpy
pandas
pyarrow
matplotlib
pyqt5
seaborn